+++ lib/src/widgets/root/root.dart
@@ -14,14 +14,15 @@ List<GlobalKey<State>> _scaffoldKeyList = <GlobalKey<State>>[];
 List<OverlayEntryMap> _overlayEntryList = <OverlayEntryMap>[];
 OverlayState _overlay;
 
-///  GlobalMaterial
-class GlobalMaterial extends StatelessWidget {
-  GlobalMaterial({
+///  GlobalWidgetsApp
+class GlobalWidgetsApp extends StatelessWidget {
+  GlobalWidgetsApp({
     Key key,
     Map<String, WidgetBuilder> routes,
     String title,
     ThemeMode themeMode,
     Locale locale,
+    Color color,
     Iterable<LocalizationsDelegate<dynamic>> localizationsDelegates,
     Iterable<Locale> supportedLocales,
     bool debugShowMaterialGrid,
@@ -38,7 +39,6 @@ class GlobalMaterial extends StatelessWidget {
     this.onUnknownRoute,
     this.builder,
     this.onGenerateTitle,
-    this.color,
     this.theme,
     this.darkTheme,
     this.localeListResolutionCallback,
@@ -46,6 +46,9 @@ class GlobalMaterial extends StatelessWidget {
     this.shortcuts,
     this.actions,
     this.onGenerateInitialRoutes,
+    this.inspectorSelectButtonBuilder,
+    this.widgetMode,
+    this.cupertinoTheme,
   })  : debugShowMaterialGrid = debugShowMaterialGrid ?? false,
         showPerformanceOverlay = showPerformanceOverlay ?? false,
         checkerboardRasterCacheImages = checkerboardRasterCacheImages ?? false,
@@ -54,6 +57,7 @@ class GlobalMaterial extends StatelessWidget {
         debugShowCheckedModeBanner = debugShowCheckedModeBanner ?? false,
         themeMode = themeMode ?? ThemeMode.system,
         title = title ?? '',
+        color = color ?? getColors(white),
         routes = routes ?? const <String, WidgetBuilder>{},
         navigatorObservers = navigatorObservers ?? <NavigatorObserver>[],
         locale = locale ?? const Locale('zh'),
@@ -67,6 +71,9 @@ class GlobalMaterial extends StatelessWidget {
               GlobalWidgetsLocalizations.delegate,
             ],
         super(key: key);
+
+  ///  风格
+  final WidgetMode widgetMode;
 
   ///  导航键
   final GlobalKey<NavigatorState> navigatorKey;
@@ -96,6 +103,9 @@ class GlobalMaterial extends StatelessWidget {
 
   ///  主题
   final ThemeData theme;
+
+  ///  Cupertino主题
+  final CupertinoThemeData cupertinoTheme;
 
   final ThemeData darkTheme;
 
@@ -140,18 +150,31 @@ class GlobalMaterial extends StatelessWidget {
   final Map<LogicalKeySet, Intent> shortcuts;
   final Map<Type, Action<Intent>> actions;
   final InitialRouteListFactory onGenerateInitialRoutes;
+  final InspectorSelectButtonBuilder inspectorSelectButtonBuilder;
 
   @override
   Widget build(BuildContext context) {
     if (navigatorKey != null) _globalNavigatorKey = navigatorKey;
-    _scaffoldKeyList = <GlobalKey<State>>[];
-    _overlayEntryList = <OverlayEntryMap>[];
-    return MaterialApp(
+    if (theme != null ||
+        debugShowMaterialGrid != null ||
+        darkTheme != null ||
+        themeMode != null ||
+        widgetMode == WidgetMode.material) materialApp();
+    if (cupertinoTheme != null || widgetMode == WidgetMode.cupertino)
+      return cupertinoApp();
+    return WidgetsApp(
         key: key,
         navigatorKey: _globalNavigatorKey,
         home: home,
         routes: routes,
         initialRoute: initialRoute,
+        pageRouteBuilder: <T>(RouteSettings settings, WidgetBuilder builder) {
+          if (widgetMode == WidgetMode.cupertino) {
+            return CupertinoPageRoute<T>(settings: settings, builder: builder);
+          } else {
+            return MaterialPageRoute<T>(settings: settings, builder: builder);
+          }
+        },
         onGenerateRoute: onGenerateRoute,
         onGenerateInitialRoutes: onGenerateInitialRoutes,
         onUnknownRoute: onUnknownRoute,
@@ -160,15 +183,12 @@ class GlobalMaterial extends StatelessWidget {
         title: title,
         onGenerateTitle: onGenerateTitle,
         color: color,
-        theme: theme,
-        darkTheme: darkTheme,
-        themeMode: themeMode,
+        inspectorSelectButtonBuilder: inspectorSelectButtonBuilder,
         locale: locale,
         localizationsDelegates: localizationsDelegates,
         localeListResolutionCallback: localeListResolutionCallback,
         localeResolutionCallback: localeResolutionCallback,
         supportedLocales: supportedLocales,
-        debugShowMaterialGrid: debugShowMaterialGrid,
         showPerformanceOverlay: showPerformanceOverlay,
         checkerboardRasterCacheImages: checkerboardRasterCacheImages,
         checkerboardOffscreenLayers: checkerboardOffscreenLayers,
@@ -177,106 +197,65 @@ class GlobalMaterial extends StatelessWidget {
         shortcuts: shortcuts,
         actions: actions);
   }
-}
 
-///  GlobalCupertino
-class GlobalCupertino extends StatelessWidget {
-  GlobalCupertino({
-    Key key,
-    Iterable<Locale> supportedLocales,
-    Locale locale,
-    Iterable<LocalizationsDelegate<dynamic>> localizationsDelegates,
-    Map<String, WidgetBuilder> routes,
-    List<NavigatorObserver> navigatorObservers,
-    String title,
-    bool showPerformanceOverlay,
-    bool checkerboardRasterCacheImages,
-    bool checkerboardOffscreenLayers,
-    bool showSemanticsDebugger,
-    bool debugShowCheckedModeBanner,
-    this.navigatorKey,
-    this.home,
-    this.theme,
-    this.initialRoute,
-    this.onGenerateRoute,
-    this.onUnknownRoute,
-    this.builder,
-    this.onGenerateTitle,
-    this.color,
-    this.localeListResolutionCallback,
-    this.localeResolutionCallback,
-  })  : showPerformanceOverlay = showPerformanceOverlay ?? false,
-        checkerboardRasterCacheImages = checkerboardRasterCacheImages ?? false,
-        checkerboardOffscreenLayers = checkerboardOffscreenLayers ?? false,
-        showSemanticsDebugger = showSemanticsDebugger ?? false,
-        debugShowCheckedModeBanner = debugShowCheckedModeBanner ?? false,
-        title = title ?? '',
-        routes = routes ?? const <String, WidgetBuilder>{},
-        navigatorObservers = navigatorObservers ?? <NavigatorObserver>[],
-        locale = locale ?? const Locale('zh', 'CN'),
-        supportedLocales = supportedLocales ??
-            <Locale>[const Locale('zh', 'CN'), const Locale('en', 'US')],
-        localizationsDelegates = localizationsDelegates ??
-            <LocalizationsDelegate<dynamic>>[
-              GlobalCupertinoLocalizations.delegate,
-              GlobalMaterialLocalizations.delegate,
-              GlobalWidgetsLocalizations.delegate
-            ],
-        super(key: key);
+  Widget materialApp() => MaterialApp(
+      key: key,
+      navigatorKey: _globalNavigatorKey,
+      home: home,
+      routes: routes,
+      initialRoute: initialRoute,
+      debugShowMaterialGrid: debugShowMaterialGrid,
+      onGenerateRoute: onGenerateRoute,
+      onGenerateInitialRoutes: onGenerateInitialRoutes,
+      onUnknownRoute: onUnknownRoute,
+      navigatorObservers: navigatorObservers,
+      builder: builder,
+      title: title,
+      onGenerateTitle: onGenerateTitle,
+      color: color,
+      theme: theme,
+      darkTheme: darkTheme,
+      themeMode: themeMode,
+      locale: locale,
+      localizationsDelegates: localizationsDelegates,
+      localeListResolutionCallback: localeListResolutionCallback,
+      localeResolutionCallback: localeResolutionCallback,
+      supportedLocales: supportedLocales,
+      showPerformanceOverlay: showPerformanceOverlay,
+      checkerboardRasterCacheImages: checkerboardRasterCacheImages,
+      checkerboardOffscreenLayers: checkerboardOffscreenLayers,
+      showSemanticsDebugger: showSemanticsDebugger,
+      debugShowCheckedModeBanner: debugShowCheckedModeBanner,
+      shortcuts: shortcuts,
+      actions: actions);
 
-  final Iterable<Locale> supportedLocales;
-  final Locale locale;
-  final Iterable<LocalizationsDelegate<dynamic>> localizationsDelegates;
-  final List<NavigatorObserver> navigatorObservers;
-  final Map<String, WidgetBuilder> routes;
-  final String title;
-  final bool showPerformanceOverlay;
-  final bool checkerboardRasterCacheImages;
-  final bool checkerboardOffscreenLayers;
-  final bool showSemanticsDebugger;
-  final bool debugShowCheckedModeBanner;
-  final GlobalKey<NavigatorState> navigatorKey;
-  final Widget home;
-  final CupertinoThemeData theme;
-  final String initialRoute;
-  final RouteFactory onGenerateRoute;
-  final RouteFactory onUnknownRoute;
-  final TransitionBuilder builder;
-  final GenerateAppTitle onGenerateTitle;
-  final Color color;
-  final LocaleListResolutionCallback localeListResolutionCallback;
-  final LocaleResolutionCallback localeResolutionCallback;
-
-  @override
-  Widget build(BuildContext context) {
-    if (navigatorKey != null) _globalNavigatorKey = navigatorKey;
-    _scaffoldKeyList = <GlobalKey<State>>[];
-    _overlayEntryList = <OverlayEntryMap>[];
-    return CupertinoApp(
-        key: key,
-        navigatorKey: _globalNavigatorKey,
-        home: home,
-        theme: theme,
-        routes: routes,
-        initialRoute: initialRoute,
-        onGenerateRoute: onGenerateRoute,
-        onUnknownRoute: onUnknownRoute,
-        navigatorObservers: navigatorObservers,
-        builder: builder,
-        title: title,
-        onGenerateTitle: onGenerateTitle,
-        color: color,
-        locale: locale,
-        localizationsDelegates: localizationsDelegates,
-        localeListResolutionCallback: localeListResolutionCallback,
-        localeResolutionCallback: localeResolutionCallback,
-        supportedLocales: supportedLocales,
-        showPerformanceOverlay: showPerformanceOverlay,
-        checkerboardRasterCacheImages: checkerboardRasterCacheImages,
-        checkerboardOffscreenLayers: checkerboardOffscreenLayers,
-        showSemanticsDebugger: showSemanticsDebugger,
-        debugShowCheckedModeBanner: debugShowCheckedModeBanner);
-  }
+  Widget cupertinoApp() => CupertinoApp(
+      key: key,
+      navigatorKey: _globalNavigatorKey,
+      home: home,
+      routes: routes,
+      initialRoute: initialRoute,
+      onGenerateRoute: onGenerateRoute,
+      onGenerateInitialRoutes: onGenerateInitialRoutes,
+      onUnknownRoute: onUnknownRoute,
+      navigatorObservers: navigatorObservers,
+      builder: builder,
+      title: title,
+      onGenerateTitle: onGenerateTitle,
+      color: color,
+      theme: cupertinoTheme,
+      locale: locale,
+      localizationsDelegates: localizationsDelegates,
+      localeListResolutionCallback: localeListResolutionCallback,
+      localeResolutionCallback: localeResolutionCallback,
+      supportedLocales: supportedLocales,
+      showPerformanceOverlay: showPerformanceOverlay,
+      checkerboardRasterCacheImages: checkerboardRasterCacheImages,
+      checkerboardOffscreenLayers: checkerboardOffscreenLayers,
+      showSemanticsDebugger: showSemanticsDebugger,
+      debugShowCheckedModeBanner: debugShowCheckedModeBanner,
+      shortcuts: shortcuts,
+      actions: actions);
 }
 
 ///  OverlayScaffold
@@ -360,7 +339,7 @@ class OverlayScaffold extends StatefulWidget {
   ///  返回按键监听
   final WillPopCallback onWillPop;
 
-  /// 刷新组件相关
+  ///  刷新组件相关
   final RefreshController controller;
   final VoidCallback onRefresh;
   final Widget child;
@@ -368,7 +347,7 @@ class OverlayScaffold extends StatefulWidget {
   final Widget footer;
   final TextStyle footerTextStyle;
 
-  /// Scaffold相关属性
+  ///  Scaffold相关属性
   final Widget bottomNavigationBar;
   final Widget appBar;
   final double appBarHeight;
@@ -389,7 +368,7 @@ class OverlayScaffold extends StatefulWidget {
   _OverlayScaffoldState createState() => _OverlayScaffoldState();
 }
 
-bool scaffoldWillPop = true;
+bool _scaffoldWillPop = true;
 
 class _OverlayScaffoldState extends State<OverlayScaffold> {
   GlobalKey<State> _globalKey = GlobalKey();
@@ -431,7 +410,7 @@ class _OverlayScaffoldState extends State<OverlayScaffold> {
   }
 
   Future<bool> onWillPop() async {
-    if (!scaffoldWillPop) return scaffoldWillPop;
+    if (!_scaffoldWillPop) return _scaffoldWillPop;
     if (widget.onWillPopOverlayClose &&
         _overlayEntryList.isNotEmpty &&
         !_overlayEntryList.last.isAutomaticOff) {
@@ -489,7 +468,9 @@ class _OverlayScaffoldState extends State<OverlayScaffold> {
   }
 }
 
-///  ************ 以下为 路由跳转 *****************///
+///  ************ 以下为 路由跳转 *****************  ///
+///
+///  打开新页面
 Future<dynamic> push(
     {WidgetBuilder builder,
     Widget widget,
@@ -497,17 +478,18 @@ Future<dynamic> push(
     RouteSettings settings,
     bool maintainState,
     bool fullscreenDialog,
-    PushMode pushMode}) {
-  return Navigator.of(_globalNavigatorKey.currentContext).push(_pageRoute(
+    WidgetMode widgetMode}) {
+  return _globalNavigatorKey.currentState.push(_pageRoute(
       title: title,
       maintainState: maintainState,
       fullscreenDialog: fullscreenDialog,
       settings: settings,
       builder: builder,
-      pushMode: pushMode,
+      pushMode: widgetMode,
       widget: widget));
 }
 
+/// 打开新页面替换当前页面
 Future<dynamic> pushReplacement(
     {WidgetBuilder builder,
     Widget widget,
@@ -515,18 +497,18 @@ Future<dynamic> pushReplacement(
     RouteSettings settings,
     bool maintainState,
     bool fullscreenDialog,
-    PushMode pushMode}) {
-  return Navigator.of(_globalNavigatorKey.currentContext).pushReplacement(
-      _pageRoute(
-          title: title,
-          maintainState: maintainState,
-          fullscreenDialog: fullscreenDialog,
-          settings: settings,
-          builder: builder,
-          pushMode: pushMode,
-          widget: widget));
+    WidgetMode widgetMode}) {
+  return _globalNavigatorKey.currentState.pushReplacement(_pageRoute(
+      title: title,
+      maintainState: maintainState,
+      fullscreenDialog: fullscreenDialog,
+      settings: settings,
+      builder: builder,
+      pushMode: widgetMode,
+      widget: widget));
 }
 
+/// 打开新页面 并移出堆栈所有页面
 Future<dynamic> pushAndRemoveUntil(
     {WidgetBuilder builder,
     Widget widget,
@@ -534,25 +516,34 @@ Future<dynamic> pushAndRemoveUntil(
     RouteSettings settings,
     bool maintainState,
     bool fullscreenDialog,
-    PushMode pushMode}) {
-  return Navigator.of(_globalNavigatorKey.currentContext).pushAndRemoveUntil(
+    WidgetMode widgetMode}) {
+  return _globalNavigatorKey.currentState.pushAndRemoveUntil(
       _pageRoute(
           title: title,
           maintainState: maintainState,
           fullscreenDialog: fullscreenDialog,
           settings: settings,
           builder: builder,
-          pushMode: pushMode,
+          pushMode: widgetMode,
           widget: widget),
       (_) => false);
 }
 
+/// 可能返回到上一个页面
+Future<bool> maybePop<T extends Object>([dynamic result]) =>
+    _globalNavigatorKey.currentState.maybePop<dynamic>(result);
+
+/// 返回上一个页面
 void pop<T extends Object>([dynamic result]) =>
-    Navigator.of(_globalNavigatorKey.currentContext).pop<dynamic>(result);
+    _globalNavigatorKey.currentState.pop<dynamic>(result);
 
-PushMode _pushMode;
+/// 循环pop 直到pop至指定页面
+void popUntil(RoutePredicate predicate) =>
+    _globalNavigatorKey.currentState.popUntil(predicate);
 
-void setGlobalPushMode(PushMode pushMode) => _pushMode = pushMode;
+WidgetMode _widgetMode;
+
+void setGlobalPushMode(WidgetMode widgetMode) => _widgetMode = widgetMode;
 
 Route<T> _pageRoute<T>({
   WidgetBuilder builder,
@@ -561,12 +552,11 @@ Route<T> _pageRoute<T>({
   RouteSettings settings,
   bool maintainState,
   bool fullscreenDialog,
-  PushMode pushMode,
+  WidgetMode pushMode,
 }) {
   assert(builder != null || widget != null);
-  if (pushMode == null && _pushMode != null) pushMode = _pushMode;
-  pushMode ??= PushMode.cupertino;
-  if (pushMode == PushMode.cupertino) {
+  _widgetMode = pushMode ?? WidgetMode.material;
+  if (_widgetMode == WidgetMode.cupertino) {
     return CupertinoPageRoute<T>(
         title: title,
         maintainState: maintainState ?? true,